+++ src/clj/lambdacd/execution/internal/util.clj
@@ -1,7 +1,7 @@
 (ns lambdacd.execution.internal.util
   (:require [lambdacd.event-bus :as event-bus]
-            [lambdacd.steps.result :as step-results]
-            [lambdacd.stepresults.merge-resolvers :as merge-resolvers]))
+            [lambdacd.stepresults.merge-resolvers :as merge-resolvers]
+            [lambdacd.stepresults.merge :as merge]))
 
 (defn send-step-result!! [{step-id :step-id build-number :build-number :as ctx} step-result]
   (let [payload {:build-number build-number
@@ -20,7 +20,7 @@
     args-with-old-and-new-globals))
 
 (defn merge-two-step-results [r1 r2]
-  (step-results/merge-two-step-results r1 r2 :resolvers [merge-resolvers/status-resolver
+  (merge/merge-two-step-results r1 r2 :resolvers [merge-resolvers/status-resolver
                                                          merge-resolvers/merge-nested-maps-resolver
                                                          merge-resolvers/combine-to-list-resolver
                                                          merge-resolvers/second-wins-resolver]))