+++ CHANGELOG.md
@@ -24,9 +24,10 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/) and this p
   * `lambdacd.steps.shell/kill`
   
 * `lambdacd.presentation.unified/unified-presentation` is unused and therefore deprecated and will be removed in the future. Use `pipeline-structure-with-step-results` instead.
-* A couple of functions were moved to make the package structure more clear: 
-  * `lambdacd.steps.result/flatten-step-result-outputs` moved to `lambdacd.stepresults.flatten/flatten-step-result-outputs`
-  * `lambdacd.steps.result/merge-step-results` moved to `lambdacd.stepresults.merge/merge-step-results`
+* A couple of functions were moved to make the package structure more clear. The existing functions still exist but are now deprecated and will be removed in the future.
+  * `lambdacd.steps.result/flatten-step-result-outputs` moved to `lambdacd.stepresults.flatten`
+  * `lambdacd.steps.result/{merge-step-results,merge-two-step-results}` moved to `lambdacd.stepresults.merge`
+  * `lambdacd.steps.result/*-resolver` moved to `lambdacd.stepresults.merge-resolvers`
   
 
 ## 0.13.0